Move window input routing context
